refactor: move PersistenceModule into di package
Aligns with DispatchersModule placement and NIA/Pokedex convention
where all Hilt modules live under a dedicated `di` subpackage.
